requirements.txt
boto3~=1.34.82
botocore~=1.34.82
openai~=1.34.0
numpy~=2.0.0
pillow~=10.3.0
moviepy~=1.0.3